Move Button from dialog to parent
Ticket: #247
